powerbi-ai-app/requirement.txt
streamlit
openai
python-docx
pymupdf